requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
openai==1.3.5
pyairtable==2.2.0
slack-sdk==3.26.0
pyngrok==7.0.0 